TranscodingStreams.jl

TranscodingStreams.jl

TranscodingStreams.jl is a package for transcoding (e.g. compression) data streams. This package exports a type TranscodingStream, which is a subtype of IO and supports various I/O operations like other usual I/O streams in the standard library.

Introduction

TranscodingStream has two type parameters, C<:Codec and S<:IO, and hence the actual type should be written as TranscodingStream{C<:Codec,S<:IO}. This type wraps an underlying I/O stream S by a codec C. The codec defines transformation (or transcoding) of the stream. For example, when C is a lossless decompression type and S is a file, TranscodingStream{C,S} behaves like a data stream that incrementally decompresses data from the file.

Codecs are defined in other packages listed below:

Package 	Library 	Format 	Codec 	Stream 	Description

CodecZlib.jl 	zlib 	RFC1952 	GzipCompression 	GzipCompressionStream 	Compress data in gzip (.gz) format.

GzipDecompression 	GzipDecompressionStream 	Decompress data in gzip (.gz) format.

RFC1950 	ZlibCompression 	ZlibCompressionStream 	Compress data in zlib format.

ZlibDecompression 	ZlibDecompressionStream 	Decompress data in zlib format.

RFC1951 	DeflateCompression 	DeflateCompressionStream 	Compress data in deflate format.

DeflateDecompression 	DeflateDecompressionStream 	Decompress data in deflate format.

CodecBzip2.jl 	bzip2 	 	Bzip2Compression 	Bzip2CompressionStream 	Compress data in bzip2 (.bz2) format.

Bzip2Decompression 	Bzip2DecompressionStream 	Decompress data in bzip2 (.bz2) format.

CodecXz.jl 	xz 	The .xz File Format 	XzCompression 	XzCompressionStream 	Compress data in xz (.xz) format.

XzDecompression 	XzDecompressionStream 	Decompress data in xz (.xz) format.

CodecZstd.jl 	zstd 	Zstandard Compression Format 	ZstdCompression 	ZstdCompressionStream 	Compress data in zstd (.zst) format.

ZstdDecompression 	ZstdDecompressionStream 	Decompress data in zstd (.zst) format.

Install packages you need by calling Pkg.add(<package name>) in a Julia session. For example, if you want to read gzip-compressed files, call Pkg.add("CodecZlib") to use GzipDecompression or GzipDecompressionStream. By convention, codec types have a name that matches .*(Co|Deco)mpression and I/O types have a codec name with Stream suffix. All codecs are a subtype TranscodingStreams.Codec and streams are a subtype of Base.IO. An important thing is these packages depend on TranscodingStreams.jl and not vice versa. This means you can install any codec package you need without installing all codec packages. Also, if you want to define your own codec, it is totally feasible like these packages. TranscodingStreams.jl requests a codec to implement some interface functions which will be described later.

Examples

Read lines from a gzip-compressed file

The following snippet is an example of using CodecZlib.jl, which exports GzipDecompressionStream{S} as an alias of TranscodingStream{GzipDecompression,S} where S<:IO:

using CodecZlib
stream = GzipDecompressionStream(open("data.txt.gz"))
for line in eachline(stream)
    # do something...
end
close(stream)

Note that the last close call will close the file as well. Alternatively, open(<stream type>, <filepath>) do ... end syntax will close the file at the end:

using CodecZlib
open(GzipDecompressionStream, "data.txt.gz") do stream
    for line in eachline(stream)
        # do something...
    end
end

Save a data matrix with Zstd compression

Writing compressed data is easy. One thing you need to keep in mind is to call close after writing data; otherwise, the output file will be incomplete:

using CodecZstd
mat = randn(100, 100)
stream = ZstdCompressionStream(open("data.mat.zst", "w"))
writedlm(stream, mat)
close(stream)

Of course, open(<stream type>, ...) do ... end works well:

using CodecZstd
mat = randn(100, 100)
open(ZstdCompressionStream, "data.mat.zst", "w") do stream
    writedlm(stream, mat)
end

Explicitly finish transcoding by writing TOKEN_END

When writing data, the end of a data stream is indicated by calling close, which may write an epilogue if necessary and flush all buffered data to the underlying I/O stream. If you want to explicitly specify the end position of a stream for some reason, you can write TranscodingStreams.TOKEN_END to the transcoding stream as follows:

using CodecZstd
using TranscodingStreams
buf = IOBuffer()
stream = ZstdCompressionStream(buf)
write(stream, "foobarbaz"^100, TranscodingStreams.TOKEN_END)
flush(stream)
compressed = take!(buf)
close(stream)

Use a noop codec

Sometimes, the Noop codec, which does nothing, may be useful. The following example creates a decompression stream based on the extension of a filepath:

using CodecZlib
using CodecBzip2
using TranscodingStreams

function makestream(filepath)
    if endswith(filepath, ".gz")
        codec = GzipDecompression()
    elseif endswith(filepath, ".bz2")
        codec = Bzip2Decompression()
    else
        codec = Noop()
    end
    return TranscodingStream(codec, open(filepath))
end

makestream("data.txt.gz")
makestream("data.txt.bz2")
makestream("data.txt")

Transcode data in one shot

TranscodingStreams.jl extends the transcode function to transcode a data in one shot. transcode takes a codec object as its first argument and a data vector as its second argument:

using CodecZlib
decompressed = transcode(ZlibDecompression(), b"x\x9cKL*JLNLI\x04R\x00\x19\xf2\x04U")
String(decompressed)

API

TranscodingStreams.TranscodingStream — Method.

TranscodingStream(codec::Codec, stream::IO; bufsize::Integer=16384)

Create a transcoding stream with codec and stream.

Examples

julia> using TranscodingStreams

julia> using CodecZlib

julia> file = open(Pkg.dir("TranscodingStreams", "test", "abra.gzip"));

julia> stream = TranscodingStream(GzipDecompression(), file)
TranscodingStreams.TranscodingStream{CodecZlib.GzipDecompression,IOStream}(<state=idle>)

julia> readstring(stream)
"abracadabra"

source

Base.transcode — Method.

transcode(codec::Codec, data::Vector{UInt8})::Vector{UInt8}

Transcode data by applying codec.

Examples

julia> using CodecZlib

julia> data = Vector{UInt8}("abracadabra");

julia> compressed = transcode(ZlibCompression(), data);

julia> decompressed = transcode(ZlibDecompression(), compressed);

julia> String(decompressed)
"abracadabra"

source

TranscodingStreams.TOKEN_END — Constant.

A special token indicating the end of data.

TOKEN_END may be written to a transcoding stream like write(stream, TOKEN_END), which will terminate the current transcoding block.

Note

Call flush(stream) after write(stream, TOKEN_END) to make sure that all data are written to the underlying stream.

source

TranscodingStreams.Noop — Type.

Noop()

Create a noop codec.

Noop (no operation) is a codec that does nothing. The data read from or written to the stream are kept as-is without any modification. This is often useful as a buffered stream or an identity element of a composition of streams.

The implementations are specialized for this codec. For example, a Noop stream uses only one buffer rather than a pair of buffers, which avoids copying data between two buffers and the throughput will be larger than a naive implementation.

source

TranscodingStreams.NoopStream — Type.

NoopStream(stream::IO)

Create a noop stream.

source

TranscodingStreams.CodecIdentity.Identity — Type.

Identity()

Create an identity (no-op) codec.

source

TranscodingStreams.CodecIdentity.IdentityStream — Type.

IdentityStream(stream::IO)

Create an identity (no-op) stream.

source

Defining a new codec

TranscodingStreams.Codec — Type.

An abstract codec type.

Any codec supporting the transcoding protocol must be a subtype of this type.

Transcoding protocol

Transcoding proceeds by calling some functions in a specific way. We call this "transcoding protocol" and any codec must implement it as described below.

There are four functions for a codec to implement:

initialize: initialize the codec

finalize: finalize the codec

startproc: start processing with the codec

process: process data with the codec.

These are defined in the TranscodingStreams and a new codec type must extend these methods if necessary. Implementing a process method is mandatory but other three are optional. initialize, finalize, and startproc have a default implementation that does nothing.

Your codec type is denoted by C and its object by codec.

The initialize(codec::C)::Void method takes codec and returns nothing. This is called once and only once before starting any data processing. Therefore, you may initialize codec (e.g. allocating memory needed to process data) with this method. If initialization fails for some reason, it may throw an exception and no other methods will be called.

The finalize(codec::C)::Void method takes codec and returns nothing. This is called when and only when the transcoding stream goes to the close state (i.e. when Base.close is called). Therefore, you may finalize codec (e.g. freeing memory) with this method. If finalization fails for some reason, it may throw an exception. Even when an exception is thrown while finalizing a stream, the stream will become the close state for safety.

The startproc(codec::C, state::Symbol)::Symbol method takes codec and state, and returns a status code. This is called just before the stream starts reading or writing data. state is either :read or :write and then the stream starts reading or writing, respectively. The return code must be :ok if codec is ready to read or write data. Otherwise, it should be :fail and then the stream throws an exception.

The process(codec::C, input::Memory, output::Memory)::Tuple{Int,Int,Symbol} method takes codec, input and output, and returns a consumed data size, a produced data size and a status code. This is called repeatedly while processing data. The input (input) and output (output) data are a Memory object, which is a pointer to a contiguous memory region with size. You must read input data from input, transcode the bytes, and then write the output data to output. Finally you need to return the size of read data, the size of written data, and :ok status code so that the caller can know how many bytes are consumed and produced in the method. When transcoding reaches the end of a data stream, it is notified to this method by empty input. In that case, the method need to write the buffered data (if any) to output. If there is no data to write, the status code must be set to :end. The process method will be called repeatedly until it returns :end status code.

source

TranscodingStreams.initialize — Function.

initialize(codec::Codec)::Void

Initialize codec.

source

TranscodingStreams.finalize — Function.

finalize(codec::Codec)::Void

Finalize codec.

source

TranscodingStreams.startproc — Function.

startproc(codec::Codec, state::Symbol)::Symbol

Start data processing with codec of state.

source

TranscodingStreams.process — Function.

process(codec::Codec, input::Memory, output::Memory)::Tuple{Int,Int,Symbol}

Do data processing with codec.

source

TranscodingStreams.Memory — Type.

A contiguous memory.

This type works like a Vector method.

source